split packagemanager_aidl from libbinder_aidl
This filegroup contains libbinder things plus packagemanager things.
Ideally this packagemanager dependency should only be taken by the
things that need it since libbinder is loaded by very many processes,
and it may consider to dirty memory in these processes (the vtables for
AIDL interfaces). This is a preparation CL.

Bug: 183654927
Test: N/A
Change-Id: Ifd952d73d5b808ae94647278a994f0d552c71314
Signed-off-by: Mohd Faraz <androiabledroid@gmail.com>